backend/requirements.txt
python-dotenv
openai
numpy
tensorflow
pillow